treble: Initial Implementation of checkhealth Command
Problem

It's good practice for a Neovim plugin to implement a plugin-specific
`checkhealth` command and treble currently does not have one.

Solution

Implement a `health` module for treble, which is what the `checkhealth` command
uses.

Result

Running the `checkhealth` command for treble, i.e. `checkhealth treble`
will indicate that `telescope` and `bufferline` are installed correctly
when they can be loaded. However, if there is a problem with either
dependency, then `treble` will fail to load at all. That behavior will be fixed
in a future change.