treble: Treble Should Load Regardless of Missing Dependencies
Problem

If the `bufferline` or `telescope` plugins are missing, then `treble`
fails to load at all. That's not ideal, and instead `treble` should be
able to load initially, and just not work properly. Additionally, the
`checkhealth` command should indicate that the required dependencies are
missing.

Solution

Move the loading of the `bufferline` and `telescope` plugins to inside
the `buffers` function instead of being part of the initial plugin
loading.

Result

`treble` will load regardless of whether the `bufferline` and
`telescope` plugins are missing. Additional work should be done to
improve the display of the `checkhealth` command when they are missing
though.