treble: Buffers without a Name Should Display as "[No Name]"
Problem

When a buffer does not have a name, which may happen on initial startup,
`bufferline` displays the buffer name as "[No Name]", whereas `treble`
displays the name as "". These should match, and they don't.

Solution

Modify `treble` shuch that buffers without a name display as "[No Name]"
and not "".

Result

`bufferline` and `treble` display a consistent name for buffers without
a name.