treble: Fix HealthCheck to Report Errors and Not Cause Errors
Problem

If either of the required dependencies for `treble`, `bufferline` or
`telescope`, are missing, then running the *checkhealth* command for
`treble` will cause an error. It displays a multiple line long trace of
errors that somewhere in the middle shows that a particular module is
missing. This is less than ideal.

Solution

For the *checkhealth* command, change loading of the required
dependencies to load via `pcall` instead of loading via `require`
directly. This allows for the ability to check for problems instead of
them being immediately reported.

Result

The *checkhealth* command for `treble` is able to show in one line for
each required dependency whether that dependency is available or not.